refactored drivetrain to subsystems folder
